[llvm-mca] Move namespace mca inside llvm::
Summary:
This allows to remove `using namespace llvm;` in those *.cpp files and simplifies the transition from the tools/mca to llvm/lib.

I decide not to have both llvm::mca:: mca:: as it will cause namespace
lookup ambiguity. The driver part can be moved to a nested namespace of llvm::mca:: when appropriate.

Reviewers: andreadb, mattd

Subscribers: javed.absar, tschuett, gbedwell, llvm-commits

Differential Revision: https://reviews.llvm.org/D53407