[IR] Disallow llvm.global_ctors and llvm.global_dtors of the 2-field form in textual format
Summary:
The 3-field form was introduced by D3499 in 2014 and the legacy 2-field
form was planned to be removed in LLVM 4.0

For the textual format, this patch migrates the existing 2-field form to
use the 3-field form and deletes the compatibility code.
test/Verifier/global-ctors-2.ll checks we have a friendly error message.

For bitcode, lib/IR/AutoUpgrade UpgradeGlobalVariables will upgrade the
2-field form (add i8* null as the third field).

Reviewers: rnk, nicholas, nlewycky, compnerd, steven_wu, dexonsmith

Reviewed By: rnk, dexonsmith

Subscribers: jyknight, mehdi_amini, sdardis, nemanjai, javed.absar, eraman, fedor.sergeev, asb, rbar, johnrusso, simoncook, apazos, sabuasal, niosHD, jrtc27, zzheng, edward-jones, atanasyan, steven_wu, rogfer01, MartinMosbeck, brucehoult, the_o, dexonsmith, PkmX, jocewei, jsji, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D61547